# storage/api-docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API for the EasyLearn app. It's a projekt done for TEKO Olten.\n\nChange the host to to 127.0.0.1:8000 to target the local development environment and dipl-backend.scapp.io to target the production system.\n\nClient ID: ZZv4TZ1aXXZaJ4B1g5rhy6EhiCDNTDrC",
    "version": "1.0.0",
    "title": "EasyLearn API",
    "contact": {
      "email": "david.weber@edu.teko.ch"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "127.0.0.1:8000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "oauth": {
      "type": "oauth2",
      "authorizationUrl": "https://dipl.eu.auth0.com/authorize?audience=http://127.0.0.1:8000",
      "flow": "implicit",
      "scopes": {
        "openid": "Standard Scope",
        "profile": "Get the users profile",
        "email": "Get the users email"
      }
    }
  },
  "security": [
    {
      "api_key": []
    },
    {
      "oauth": [
        "openid",
        "profile"
      ]
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Benutzer"
        ],
        "summary": "Alle Benutzer",
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Benutzer"
        ],
        "summary": "Neuen Benutzer hinzufügen",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserInvite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
...
    "/subjects": {
      "get": {
        "tags": [
          "Fächer"
        ],
        "summary": "Alle Fächer",
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Subject"
              }
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Fächer"
        ],
        "summary": "Neues Fach hinzufügen",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SubjectPut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "schema": {
              "$ref": "#/definitions/Subject"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
...
    "/subjects/{id}/lessons": {
      "get": {
        "tags": [
          "Fächer",
          "Unterricht"
        ],
        "summary": "Der Unterricht eines Fachs",
        "parameters": [
          {
            "$ref": "#/parameters/ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Lesson"
              }
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Fächer",
          "Unterricht"
        ],
        "summary": "Neuer Unterricht für ein Fach hinzufügen",
        "parameters": [
          {
            "$ref": "#/parameters/ID"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LessonPut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "schema": {
              "$ref": "#/definitions/Lesson"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
...
    "/subjects/{id}/users": {
      "get": {
        "tags": [
          "Fächer",
          "Benutzer"
        ],
        "summary": "Die Benutzer eines Fachs",
        "parameters": [
          {
            "$ref": "#/parameters/ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Erfolgreich",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Timestamps": {
      "type": "object",
      "properties": {
        "createAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "IdentifyAble": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        }
      }
    },
    "User": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/IdentifyAble"
        },
        {
          "$ref": "#/definitions/UserPut"
        },
        {
          "type": "object",
          "properties": {
            "calenderToken": {
              "type": "string"
            }
          }
        },
        {
          "$ref": "#/definitions/Timestamps"
        }
      ]
    },
    "UserPut": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": [
            "student",
            "teacher",
            "admin"
          ]
        }
      }
    },
    "UserInvite": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/UserPut"
        },
        {
          "type": "object",
          "required": [
            "inviteEmail"
          ],
          "properties": {
            "inviteEmail": {
              "type": "string"
            }
          }
        }
      ]
    },
...
    "Subject": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/IdentifyAble"
        },
        {
          "$ref": "#/definitions/SubjectPut"
        },
        {
          "type": "object",
          "properties": {
            "status": {
              "type": "string",
              "enum": ["done", "notask", "open"]
            }
          }
        },
        {
          "type": "object",
          "properties": {
            "teacher": {
              "$ref": "#/definitions/User"
            }
          }
        },
        {
          "$ref": "#/definitions/Timestamps"
        }
      ]
    },
    "SubjectPut": {
      "type": "object",
      "required": [
        "name",
        "archived",
        "teacherId"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "archived": {
          "type": "boolean",
          "default": false
        },
        "teacherId": {
          "type": "integer"
        }
      }
    },
...
    "Lesson": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/IdentifyAble"
        },
        {
          "$ref": "#/definitions/LessonPut"
        },
        {
          "type": "object",
          "properties": {
            "subjectId": {
              "type": "integer"
            },
            "status": {
              "type": "string",
              "enum": ["done", "notask", "open"]
            }
          }
        },
        {
          "$ref": "#/definitions/Timestamps"
        }
      ]
    },
    "LessonPut": {
      "type": "object",
      "required": [
        "startDate",
        "endDate",
        "canceled"
      ],
      "properties": {
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "location": {
          "type": "string"
        },
        "room": {
          "type": "string"
        },
        "canceled": {
          "type": "boolean",
          "default": false
        },
        "type": {
          "type": "string",
          "enum": [
            "lesson",
            "exam",
            "reminder"
          ]
        }
      }
    },
...
  },
  "parameters": {
    "ID": {
      "name": "id",
      "in": "path",
      "description": "ID der ersten Ressource",
      "required": true,
      "type": "integer",
      "format": "int64"
    },
...
  },
  "responses": {
    "NotFound": {
      "description": "Die Ressource konnt nicht gefunden werden"
    },
...
    "ServerError": {
      "description": "Server Fehler"
    },
    "ValidationError": {
      "description": "Die Ressource ist nicht valid"
    },
    "Unauthorized": {
      "description": "Sie sind nicht autorisiert (401)."
    },
...
  }
}